+++ src/__tests__/CountUp.test.js
